155 restructure generate table structure (#156)
* restructure


* rename analyze to finder


---------
