+++ src/main/java/de/uol/pgdoener/th1/business/infrastructure/analyzeTable/core/CellInfo.java
@@ -1,4 +1,4 @@
-package de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core;
+package de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core;
 
 /**
  * Represents a cell in the matrixInfo, including its column ID and whether it contains data.
