+++ src/main/java/de/uol/pgdoener/th1/business/infrastructure/analyzeTable/core/MatrixInfoService.java
@@ -1,4 +1,4 @@
-package de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core;
+package de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
