+++ src/main/java/de/uol/pgdoener/th1/business/infrastructure/analyzeTable/core/ValueType.java
@@ -1,0 +1,8 @@
+package de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core;
+
+public enum ValueType {
+    NUMBER,
+    BOOLEAN,
+    STRING,
+    EMPTY
+}
