+++ src/main/java/de/uol/pgdoener/th1/business/infrastructure/analyzeTable/finder/FindEmptyColumnService.java
@@ -1,9 +1,9 @@
-package de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.analyze;
+package de.uol.pgdoener.th1.business.infrastructure.analyzeTable.finder;
 
 import de.uol.pgdoener.th1.business.dto.EmptyColumnReportDto;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.ColumnInfo;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.ColumnInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.MatrixInfo;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.ColumnInfo;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.ColumnInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.MatrixInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
