--- src/main/java/de/uol/pgdoener/th1/business/infrastructure/generatetablestructure/core/ValueType.java
@@ -1,8 +1,0 @@
-package de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core;
-
-public enum ValueType {
-    NUMBER,
-    BOOLEAN,
-    STRING,
-    EMPTY
-}
