+++ src/test/java/de/uol/pgdoener/th1/business/infrastructure/analyzeTable/factory/CellInfoFactoryTest.java
@@ -1,7 +1,7 @@
-package de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.factory;
+package de.uol.pgdoener.th1.business.infrastructure.analyzeTable.factory;
 
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.CellInfo;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.ValueType;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.CellInfo;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.ValueType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
