+++ src/test/java/de/uol/pgdoener/th1/business/service/GenerateTableStructureServiceTest.java
@@ -2,14 +2,14 @@ package de.uol.pgdoener.th1.business.service;
 
 import de.uol.pgdoener.th1.business.dto.*;
 import de.uol.pgdoener.th1.business.infrastructure.InputFile;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.AnalyzeMatrixInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.analyze.*;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.CellInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.ColumnInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.MatrixInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.core.RowInfoService;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.factory.CellInfoFactory;
-import de.uol.pgdoener.th1.business.infrastructure.generatetablestructure.factory.MatrixInfoFactory;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.AnalyzeMatrixInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.CellInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.ColumnInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.MatrixInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.core.RowInfoService;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.factory.CellInfoFactory;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.factory.MatrixInfoFactory;
+import de.uol.pgdoener.th1.business.infrastructure.analyzeTable.finder.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
